backend/requirements.txt
# --- Core Flask Framework ---
Flask==3.0.3
Flask-Cors==4.0.1

# --- Database and ORM ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30

# --- Schema Serialization ---
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.2

# --- Database Migrations ---
Flask-Migrate==4.0.7
alembic==1.13.2

# --- Authentication (optional, if using JWT later) ---
Flask-JWT-Extended==4.6.0

# --- Utilities ---
python-dotenv==1.0.1

# --- Testing (optional) ---
pytest==8.3.2

# --- For Development ---
black==24.4.2

# --- External API Integrations ---
requests==2.32.3
openai==1.3.0
python-dotenv==1.0.1
